Create Number_Guessing_Game.py
This is a number guessing game in which computer will think a random number between 1 to 100 and then user will have to guess that number in given numbers of attempt. User can select between 2 levels, easy or hard. In easy user will get 10 attempts and in hard user will get 5 attempts to guess the correct number.